+++ docs/pages/releases.md
@@ -17,11 +17,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- 新增 navbar.animated 配置项，可以进行配置导航栏动画
+- 增加 `navbar.animated` 配置项，支持导航栏动画效果的自定义配置
+- 增加 `translateIcon` 和 `toggleThemeIcon` 配置项，允许用户自定义相应的按钮图标
 
 ### Changed
 
 - 调整导航栏的样式
+- SCSS 命名空间由 `st` 变更为 `sakura`
 
 ## [0.6.8] - 2024-10-02